+++ src/app/form/form-reactive-form.demo.html
@@ -68,6 +68,26 @@
         </div>
       </div>
 
+      <div class="field">
+        <label soho-label>Radio Button</label>
+        <input soho-radiobutton (change)="onChange($event)" id="optionVWO1" type="radio" name="radiobutton" value="1" checked formControlName="radiobutton"/>
+        <label soho-label for="optionVWO1" [forRadioButton]="true">One</label>
+        <br>
+        <input soho-radiobutton (change)="onChange($event)" id="optionVWO2" type="radio" name="radiobutton" value="2" formControlName="radiobutton"/>
+        <label soho-label for="optionVWO2" [forRadioButton]="true">Two</label>
+        <br>
+        <input soho-radiobutton (change)="onChange($event)"   id="optionVWO3" type="radio" name="radiobutton" value="3" formControlName="radiobutton"/>
+        <label soho-label for="optionVWO3" [forRadioButton]="true">Three</label>
+        <br>
+        <input soho-radiobutton (change)="onChange($event)" [disabled]="isDisabled" id="optionVWO4" type="radio" name="radiobutton" value="4" formControlName="radiobutton"/>
+        <label soho-label for="optionVWO4" [forRadioButton]="true">Four Disabled</label>
+      </div>
+
+      <div class="field">
+        <label for="textarea" class="label">Textarea</label>
+        <textarea soho-textarea maxlength="90" name="textarea" formControlName="textarea"></textarea>
+      </div>
+
       <button soho-button="tertiary" (click)="toggleDisableForm()">Toggle Disable Form</button>
 
       <h3>Form valid?: {{ demoForm.status | json }}</h3>
@@ -81,6 +101,7 @@
       <h3>Autocomplete Value: {{ demoForm.controls.autocomplete.value | json }}</h3>
       <h3>Slider Value: {{ demoForm.controls.slider.value | json }}</h3>
       <h3>Lookup Value: {{ demoForm.controls.lookup.value | json }}</h3>
+      <h3>RadioButton Value: {{ demoForm.controls.radiobutton.value | json }}</h3>
     </div>
   </form>
 </div>
