+++ src/soho/dropdown/soho-dropdown.spec.ts
@@ -5,10 +5,11 @@ import { FormsModule, FormGroup, FormBuilder } from '@angular/forms';
 
 import { SohoDropDownModule } from './soho-dropdown.module';
 import { SohoDropDownComponent } from './soho-dropdown.component';
+import { ReactiveFormsModule } from '@angular/forms';
 
 @Component({
   template: `
-  <select soho-dropdown noSearch  [closeOnSelect]="true" [(ngModel)]="selectedOption">
+  <select soho-dropdown noSearch [closeOnSelect]="true" [(ngModel)]="selectedOption">
     <option *ngFor="let option of options" [value]="option.value">{{option.label}}</option>
   </select>`
 })
@@ -26,43 +27,6 @@ class SohoDropDownTestComponent {
     { value: 'WA', label: 'Washington' },
     { value: 'WY', label: 'Wyoming' }
   ];
-}
-
-@Component({
-  template: `
-  <form>
-    <select soho-dropdown formControlName="state>
-      <option *ngFor="let state of states" [value]="state.value">{{state.label}}</option>
-    </select>
-  </form>`
-})
-class SohoDropDownReactiveFormTestComponent {
-  @ViewChild(SohoDropDownComponent) dropdown: SohoDropDownComponent;
-
-  public options = [
-    { value: 'AK', label: 'Alaska' },
-    { value: 'AZ', label: 'Arizona' },
-    { value: 'CA', label: 'California' },
-    { value: 'CO', label: 'Colorado' },
-    { value: 'MN', label: 'Minnesota' },
-    { value: 'ND', label: 'North Dakota' },
-    { value: 'OR', label: 'Oregon' },
-    { value: 'WA', label: 'Washington' },
-    { value: 'WY', label: 'Wyoming' }
-  ];
-
-  formGroup: FormGroup;
-
-  constructor(private formBuilder: FormBuilder) {
-    this.formGroup = this.createForm();
-    this.formGroup.disable();
-  }
-
-  private createForm() {
-    return this.formBuilder.group({
-      state: 'ND'
-    });
-  }
 }
 
 describe('Soho Dropdown Unit Tests', () => {
@@ -198,41 +162,5 @@ describe('Soho Dropdown Render', () => {
     fixture.detectChanges();
 
     expect(el.hasAttribute('disabled')).toBeTruthy();
-  });
-});
-
-describe('Soho Dropdown Reactive Forms', () => {
-  let dropdown: SohoDropDownComponent;
-  let component: SohoDropDownReactiveFormTestComponent;
-  let fixture: ComponentFixture<SohoDropDownReactiveFormTestComponent>;
-  let de: DebugElement;
-  let el: HTMLElement;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      declarations: [SohoDropDownTestComponent],
-      imports: [FormsModule, SohoDropDownModule]
-    });
-
-    fixture = TestBed.createComponent(SohoDropDownReactiveFormTestComponent);
-    component = fixture.componentInstance;
-    dropdown = component.dropdown;
-
-    de = fixture.debugElement;
-    el = de.query(By.css('select[soho-dropdown]')).nativeElement;
-
-    fixture.detectChanges();
-  });
-
-  it('Check "disabled"', () => {
-    component.formGroup.disable();
-    fixture.detectChanges();
-
-    expect(el.hasAttribute('disabled')).toBeTruthy('disabled');
-
-    component.formGroup.enable();
-    fixture.detectChanges();
-
-    expect(el.hasAttribute('disabled')).toBeFalsy('disabled');
   });
 });
